solución de problema con input numericos para aceptar 2 dígitos y se agregó montos totales a libro de ventas.

+++ database/migrations/2020_02_15_211007_create_proformas_detalles_table.php
@@ -17,8 +17,8 @@ class CreateProformasDetallesTable extends Migration {
 			$table->increments('id');
 			$table->integer('proforma_id')->nullable();
 			$table->integer('producto_id')->nullable();
-			$table->decimal('precio', 8,2)->nullable();
-			$table->integer('cantidad')->nullable();
+			$table->decimal('precio', 10)->nullable();
+			$table->decimal('cantidad', 10)->nullable();
 			$table->timestamps();
 		});
 	}
